Move routes and controllers into seperate files
