+++ routes/tourRoutes.js
@@ -1,0 +1,17 @@
+const express = require('express');
+const tourControllers = require('./../controllers/tourControllers');
+
+const router = express.Router();
+
+router
+  .route('/')
+  .get(tourControllers.getAllTours)
+  .post(tourControllers.createTour);
+
+router
+  .route('/:id')
+  .get(tourControllers.getTour)
+  .patch(tourControllers.updateTour)
+  .delete(tourControllers.deleteTour);
+
+module.exports = router;
